optimize component structure for server-side rendering
