+++ src/app/(dashboard)/dashboard/search/layout.tsx
@@ -6,12 +6,8 @@
  * Layout for search pages with navigation tabs
  */
 
-"use client";
-
-import Link from "next/link";
-import { usePathname } from "next/navigation";
 import type { ReactNode } from "react";
-import { cn } from "@/lib/utils";
+import { SearchTabsClient } from "@/features/search";
 
 interface SearchLayoutProps {
   children: ReactNode;
@@ -24,34 +20,9 @@ const searchTabs = [
 ];
 
 export default function SearchLayout({ children }: SearchLayoutProps) {
-  const pathname = usePathname();
-
   return (
     <div className="space-y-6">
-      {/* Search Tabs */}
-      <div className="border-b border-border">
-        <nav className="-mb-px flex space-x-8">
-          {searchTabs.map((tab) => {
-            const isActive = pathname === tab.href;
-            return (
-              <Link
-                key={tab.href}
-                href={tab.href}
-                className={cn(
-                  "border-b-2 px-1 pb-4 text-sm font-medium transition-colors",
-                  isActive
-                    ? "border-primary text-primary"
-                    : "border-transparent text-muted-foreground hover:border-border hover:text-foreground",
-                )}
-              >
-                {tab.label}
-              </Link>
-            );
-          })}
-        </nav>
-      </div>
-
-      {/* Page Content */}
+      <SearchTabsClient tabs={searchTabs} />
       {children}
     </div>
   );
