+++ src/features/search/components/index.ts
@@ -2,3 +2,7 @@ export * from "./SearchInput";
 export * from "./UserSearchCard";
 export * from "./CampusSearchCard";
 export * from "./SearchResults";
+export * from "./StudentsSearchClient";
+export * from "./MentorsSearchClient";
+export * from "./CampusesSearchClient";
+export * from "./SearchTabsClient";
