+++ src/features/search/hooks/useInfiniteScroll.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { useEffect, useRef } from "react";
 
 interface UseInfiniteScrollProps {
