+++ src/features/search/hooks/useSearchMentors.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { useQuery } from "@tanstack/react-query";
 import { useState, useEffect } from "react";
 import { searchMentors } from "../api";
@@ -24,7 +26,7 @@ export function useSearchMentors(initialQuery = "") {
         pageIndex: page,
         perPage: 30,
       }),
-    enabled: debouncedQuery.length >= 2,
+    enabled: debouncedQuery.length >= 3,
     staleTime: 5 * 60 * 1000,
   });
