+++ src/features/search/hooks/useSearchUsers.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { useQuery } from "@tanstack/react-query";
 import { useState, useEffect } from "react";
 import { searchUsers } from "../api";
@@ -25,7 +27,7 @@ export function useSearchUsers(initialQuery = "") {
         pageIndex: page,
         perPage: 30,
       }),
-    enabled: debouncedQuery.length >= 2,
+    enabled: debouncedQuery.length >= 3,
     staleTime: 5 * 60 * 1000, // 5 minutes
   });
